ChatroomManager: wrote createChatroom, addUser, removeUser, sendMessage
Important follow-up notes:
sendMessage - do we want to be able to keep track of who sent what messages? how would we do this?

privateChats - currently, this is a list of all 2 person Chatrooms. This assumes all other Chatrooms are those associated with each Activity (stored in conferenceChats). What if some Users want to create a 3- or 4-person group chat? Maybe we can expand the definition of privateChats to be more flexible //OR// set a precondition that any new Chatrooms we create will only have 2 people. Those in conferenceChats are constructed during the construction of Activity objects.